Commiting Build 1.2.46

+++ source/code/debug.c
@@ -276,7 +276,7 @@ void debug_dataset(data_set_t *data){
         
         row_t *temp_row=&data->rows[i];
         printf("Length %ld\n",temp_row->column_length);
-        for(int b=0;temp_row->column_length;b++){
+        for(long b=0;temp_row->column_length;b++){
             if(b>10) break;
             if(b!=0) printf("%ld %ld,",b,temp_row->column_length);
